commit for updating and deleting user

+++ src/main/java/com/github/user_manager/repository/UserProfilesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserProfilesRepository extends JpaRepository<UserProfiles, Long> {
+public interface UserProfilesRepository extends JpaRepository<UserProfiles, Integer> {
 }
